qa_bot/qa_bot.py
# qa_bot.py
import PyPDF2
import pinecone  # Ensure you've set up Pinecone API

def process_pdf(file):
    text = ""
    reader = PyPDF2.PdfReader(file)
    for page in reader.pages:
        text += page.extract_text() + "\n"
    return text

def qa_bot(query, pdf_file):
    # Process the uploaded PDF
    pdf_text = process_pdf(pdf_file)
    
    # Here, you'd typically generate embeddings and store them in Pinecone
    # Example: embeddings = generate_embeddings(pdf_text)

    # Example response
    answer = "This is a placeholder answer based on your query."
    document_segments = "Relevant segments from the document."
    
    return answer, document_segments
