qa_bot/requirements.txt
streamlit
gradio
PyPDF2
pinecone-client
